Opening and closing/disposing the DevTools for browsers works
OK, this is a quite big commit (see info further down). I implemented
opening and closing the DevTools for browsers in this commit.

The main problem here is that we have to close browsers (also the
browsers that display the DevTools) correctly, otherwise the application
will not cleanly exit later. And browsers have a two-step close process:
First, we call close(false, runnable), and when the CEF browser calls
back to the runnable, we call setCloseAllowed() and close(true). Yikes.

Another problem I faced was that when I put multiple devtools browsers
in a single window (with a tabbed pane), *none* of them would dispose
correctly. Strangely, putting multiple browsers in the same window works
for normal browsers (see "SecondWindow"). Anyway, MetalShell now opens
all DevTools in their own window. Not as nice as a TabbedPane, but works
for now.

In more detail, the WIP commits were:
* Open new window (browser tools and dispose not working yet)
* no gpu rendering, dispose dev tools more nicely
* try to dispose the browsers in the second window more nicely
* Can correctly close the browser of a browser window.
* can open second window, but not close it correctly
* can close devtools of first window correctly, but not remove a browser from devtools or reuse devtools window
* Can open devtools in tabs of a single window, but does not close them correctly yet.
* can close DevTools and frame of first browser correctly
* closes browser and devtools correctly for first window
* added closing code for second window, but does not remove correctly yet
* closing a single devtools browser in a frame work. So, not tabbed pane possible :(
* Open DevTools window when browser was created
* last browser closed callback
* Browser can now notify others when browser creation was finished
* dev tools with frame factory
* correctly opens and disposes DevTools frames
* closes cef app correctly when last browser closes